Rewrote this test to reduce the number of "special cases" and to pass more reliably, particularly when a large number of process are running.
First, this adds two new fields to the fun_names tuple:
1) A tuple of attributes of the ntpinfo struct to compare the return
   value of the method being tested against.  There is one special
   case, currently used only for proc_memory_info, where instead of a
   tuple this can be a callable taking the PID as an argument, and
   returning a value to compare to the tested method.
2) A tuple (which may be empty) of PIDs for which the test should be
   skipped.

Next, this changes the test to suspend each process before calling
proc_info, as well as the methods to test it against, and does not
resume the process until it has been fully tested.  If it cannot
resume a process it skips it.  This reduces the possibility of race
conditions (which occurred especially for the proc_io_counters test
with probability increasing with the number of running processes).