In practice, installing the full dev-env is proving a bit troublesome.
It takes a long time; in particular to build grpcio which is broken on
Cygwin anyways, but is listed as a dependency of one of the
dependencies.

I don't think we need all this just to build psutil and run te tests, so
here's a simplified script based on the one for freebsd.