Cygwin: Add a new function for querying NT symbolic links.
In particular I need this on Cygwin to convert certain device paths
in the form \Device\HarddiskN\PartitionN to their DOS device paths
so that I can look up their associated drive letter (if any).

I included this in arch/windows/disk.c since it was related to some of
the functions already there; however since I only need it on Cygwin it
might be better to move it to _psutil_cygwin.c or some arch/cygwin/
file.  I leave this up to @giampaolo.